pyproject.toml
[tool.poetry]
name = "timelapse"
version = "0.1.0"
description = "quick and dirty project to generate a timelapse"
license = "MIT"
authors = ["AlbertUnruh <AlbertUnruh@pm.me>"]
readme = "README.md"
repository = "https://github.com/AlbertUnruh/timelapse/"
homepage = "https://github.com/AlbertUnruh/timelapse/"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydantic-settings = "^2.7.0"
pydantic = "^2.10.4"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
numpy = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py312"]
include = "\\.pyi?$"

[tool.isort]
line_length = 120
py_version = 312
profile = "black"
lines_after_imports = 2
lines_between_sections = 1
split_on_trailing_comma = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
import_heading_stdlib = "standard library"
import_heading_thirdparty = "third party"
import_heading_firstparty = "first party"
import_heading_localfolder = "local"

[tool.ruff]
line-length = 120
target-version = "py312"
output-format = "concise" # set to full if you want to see the source of the error/warning

[tool.ruff.lint]
task-tags = ["TODO", "FIXME", "NOTE"]
select = [
    "F",    # Pyflakes
    "E",    # pycodestyle
    "W",    # pycodestyle
    "C90",  # mccabe
    "N",    # pep8-naming
    "D",    # pydocstyle
    "UP",   # pyudpgrade
    "YTT",  # flake8-2020
    "ANN",  # flake8-annotations
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "FBT",  # flake8-boolean-trap
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T10",  # flake8-debugger
    "EM",   # flake8-errmsg
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
    "INP",  # flake8-no-pep420
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SLF",  # flake8-self
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "TC",   # flake8-type-checking
    "INT",  # flake8-gettext
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "PLC",  # Pylint Convention
    "PLE",  # Pylint Error
    "PLR",  # Pylint Refactor
    "PLW",  # Pylint Warning
    "TRY",  # tryceratops
    "FLY",  # flynt
    "PERF", # Perflint
    "FURB", # refurb
    "RUF",  # Ruff-specific rules
]
ignore = [
    "D100",   # Missing docstring in public module
    "D103",   # Missing docstring in public function
    "D104",   # Missing docstring in public package
    "ANN003", # Missing type annotation for `**kwargs`
    "G004",   # Logging statement uses f-string
]
per-file-ignores = {"timelapse/copper_wire.py" = ["S113"], "timelapse/__main__.py" = ["T201"]}

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true
suppress-dummy-args = true
suppress-none-returning = true

[tool.ruff.lint.flake8-errmsg]
max-string-length = 20

[tool.ruff.lint.flake8-gettext]
extend-function-names = ["get_text"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.mccabe]
max-complexity = 13

[tool.ruff.lint.pydocstyle]
convention = "numpy"
